EditDialog: extract WikimediaCommonsEditor + paste URL
